requirements.txt
streamlit==1.0.0
PIL
azure
msrests